refactor: 에러처리 메서드 파일 분리
